v0.0.2 UPDATE:
Code tested for the 2x2 Hubbard model at U=0, against tight binding results, and
U=2, against statistically exact HF-QMC results (provided by A.Valli)

The actual implementation of the spectral sum was correct only for the local
function, but miserably failed for the non-local ones (probably including
inter-orbital quantities).
A new version of the procedure is now in place giving correct results.
The central problem was that creation and destruction operator matrix
elements have to be evaluated independently in the eigenbasis of the H
decomposition (the innermost loop). This corresponds to the fact
the while the trace and unity decomposition summations are external
and act on both operators, the decomposition over the eigenbasis components
is internal, and as such independent for the two operators.