VCA_plus_BATH Development version
This version includes non-interacting, bath, sites along with the
correlated cluster, or reference system.
The implementation closely follow the DMFT_ED one, plus some check
on the presence (not mandatory here) of the bath.

>>> ACTHUNG <<<
I am not sure whether we need to change also the evaluation of the
Q and Lambda matrices. So far I think these are only for the correlated
sites, so bath remains excluded. This is similar in some sense to
the DMFT case.