v0.0.7 BUG FIX & UPDATE:
- This is a new version of the code with bath included (only normal form).
We removed a relevant bug in VCA_MAIN when evaluating the M matrix.
The code was using essentially only the first Ncluster X Ncluster block of
the Vmat matrix, screwing up the entire calculation. Now it is fixed by
extending the Qcluster matrix to tile the system.

Yet the code has problems and needs to be debugged.