v1.0.0: NEW CODE VERSION.
This is a major reshape/rewriting of the code. We move to a Green's function
based calculation and will exploit translational symmetry + periodicity.

In this step we reimplemented the GF calculation using Full Exact Diagonalization
and compared with exact QMC results for the 2x2 square lattice.
Benchmark is excellent and is identical to the initial tests of the previous code
(working in a different way).

The bath is implemented but not used. Changes are in order for this.
Evaluation of the Self-energy function is not implemented.

We shall now implment the Lanczos based algorithm (hey we're gonna copy
from DMFT ED huh :-))