v1.1.1 UPDATE
Code updated with storage of the spectral representation of the
Green's Function (GF), using Weights and Poles. This enabled a fast
re-calculation of the cluster GF on a given z\in\CCC or an array
z(:)\in\CCC, as performed in vca_gf_cluster in VCA_IO.

This now enables to evaluate the \Omega grand potential integral,
both in the Matsubara or Real-axis formalism.