v0.99 BETA VERSION
First working version of the code for the Nup_Ndw branch. The code is
now capable of returning the Omega potential for the given system and
cluster.
Tested against QMC data for a 2x2 cluster and ED data for 1d chain.

The user needs to implement the routines producing the t and t' hopping
matrices and anything problem-dependent (draw the Omega curves, minimize
Omega, periodize the Green's function and the Self-energy) in the driver.

t' is passed as a (nlat,nlat,nspin,nspin,norb,norb) object, t as a (nlat,nlat,nspin,nspin,norb,norb,Nk) object.
The code calculating Omega (VCA_OMEGA) can be ported to the main branch provided
modifications in VCA_MAIN are also made.

TODO:

- save weights&poles in file as gf_cluster_restart.vca
- make it possible to pass t as a (Nlso,Nlso,Nk) object (useful for
superconducting case, that is not realizable in this branch)
- calculate the Self-Energy
- fix flags for the integration routines (actually in Scifor)
- extend to orbital-resolved case
- add bath
- add finite T case