VERSION 1.0
VCA code version 1.0: the code supports trial systems with 1 variational
parameter at a time. Variational parameters can be 1-body params for
the trial cluster or fictitious bath parameters (on-site energy and hybridization). Only the "normal" type of bath is supported for now. In case of added bath
sites, the use of HFMODE=F is recommended.