Intermediate commit: (POTENTIALLY) CRITICAL BUG FIX
This update fixes a bug introduced in v0.12.7 (840fb6080a216f667).
In that release I switched the nullify statements to delete
Retrieved eigenvectors to actual deallocate statements.
The reason is that in the MPI execution mode the pointer used to
retrieve the eivenvector point to a temporary vector, with full
size only the master.
After use this tmp vectory must be deleted.
The nullify statements however where merely unlink the pointers
generating a malicious leak.

However the change nullify --> deallocate crash the execution in the
serial mode, where the pointer point directly to the eigenvectors.
deallocating the pointer means deleting the eigenvectors!

This bug (noted by Pablo) is fixed by using the usual pre-compilation
structure as used to retrieve the vector and which makes a
differentiation among MPI and serial execution.
SERIAL --> nullify
MPI    --> deallocate

In a next update I will rewrite this part so that the pointer can
point to the corresponding chunk of the eigenvector in any cpu.
This will save memory and simplify treatment, but will involve
more communication with Send/Recv operations, which can however be
gathered toghether or masked.