CRITICAL BUGFIX:
VCA_SETUP.f90: upon constructing getCsector and getCdgsector, the loop
has to be split into four (UP-add/delete, DOWN-add/delete), otherwise
the if condition causes it to cycle entirely if the first check fails.
This causes wrong mapping in cases in which the GS is found to be on
"edge" sectors (one index is 0 or Ns).