VERSION 1.4:
VCA_OBSERVABLES.f90: added possibility for user-defined variables.
The user has to call observables_cluster(matrix,value) where the
optional aruments are the (Nlat,Nlat,Nspin,Nspin,Norb,Norb) observable
matrix s_ij and the variable in which to store the result. If no
variable is provided the result is only printed to the logfile. If no
matrix is provided the local observables are calculated in the usual
way. Checked consistent results.