UPDATE CMAKELISTS.TXT
If we are compiling with external BLAS and LAPACK CMake first searches
for a module for MKL. If it's found it imports the libraries. If not it
searches for system BLAS and LAPACK. On some distros (e.g Debian) when
MKL packages are installed global variables are not updated but the
libraries for BLAS and LAPACK become symlinks to the MKL ones. In this
case CMake is agnostic with respect to the source of the system
libraries, it is assumed the user knows what packages he installed.