VERSION 1.5 UPDATE
The CMakeLists.txt has been updated so to support linking with
3rd party Blas/Lapack libraries for better optimization.

Blas/Lapack are searched from MKL or OS.
Note that in some circumstances MKL links directly inside the OS
so it is found as a regular blas/lapack set.
On Apple VecLib libraries from Xcode are used to get get best
performances (MKL level or even better)

Enjoy this code.