UPDATE:
Introduced the flag HFSHIFT for systems with bath: only considered if
HFMODE=TRUE, if HFSHIFT=TRUE bath local energies are defined with
respect to the half-filling chemical potential value.
That is, for example, (E=0 & HFSHIFT=T) == (E=Uloc/2 && HFSHIFT=F)
for the 1-band Hubbard model.
This is needed because the on-site energy is itself a variational
parameter and the user must be able to modify it, but determining the
expression for the half-filling value for multi-orbital systems can be
difficult.

Modified old drivers with bath to account for the previous change.

Added vca_bhz_bath driver (for now 1 site + 1 bath site)