v 1.5.8 UPDATE:
Removed global input variables NKPTS and NDIM. The VCA code only cares
about the total number of k-points, disregarding the way the mesh is
constructed. Therefore any variable related to that is problem-dependent
and should be contained in the driver.

All existing drivers have been corrected to work with the new code, the
way they initialize the mesh is by having a local input variable NKPTS,
which is an array of fixed dimension with each element representing the
number of k-points in the given direction. This also makes possible to
create asymmetrical clusters (e.g. rectangular).

Tests performed with vca_square_bath and vca_bhz_bath are consistent
with previous results.

WARNING!! previous input files do not work correctly anymore, since the
type of the input variable NKPTS has changed and NDIM is not an input
variable anymore. NKPTS should now be provided in the same way as Uloc
(with the right dimension). Check the drivers in case of errors.