v 1.5.9 UPDATE
IMPORTANT UPDATE: due to a bug in Arpack (to be fixed in SciFortran), in
some cases with big clusters or multi-orbital systems the exact
diagonalization routine is unable to find the correct ground state. This
obviously creates incoherent results for the Omega potential and all the
deriving quantitis. A separate branch of the code has been created that
uses the real-valued diagonalization routine, which is not affected by
the problem.
As an alternative, and only for ZERO temperature, a lanczos routine has
been included (it can be selected via the flag "LANC_METHOD"), which
gives the correct result.

Cosmetic improvements to the bath printout in the LOGfile have been
included.

Fixed a bug in VCA_OBSERVABLES that generated segfault for NSPIN=1