!!! CRITICAL BUGFIX !!!:
During debug of BHZ model in CDMFT we realized that we have always
constructed the transposed H_up and H_down small matrices. This had no
effect in the dobule/symmetric case. This inclues in particular the
single-site case. However, if complex hopping elements are present (as
in a BHZ cluster) the impurity matrix turns out to be the transposed
(i.e. the conjugate) of the correct one.

Needs to be ported in every other ED-based code.