Initial commit:
added option to pass different values of the interaction parameters
per each site in the cluster. These have to be passed from the driver
as optional arrays to the vca_solve routines. In this case, the input
file parameters will be overrided. If no such arrays are provided the
code defaults to the input file values. To be tested.