BUGFIX part 2:
this seems to be better: we broadcast the norm of the vector before we
do delete_hv_sector. In this way, only the cores in mpi_comm_group
receive it. Then the others will run add_to_lanczos and fail.
Now we broadcast after delete_hv_sector. Seems to be fixed, waiting
confirmation to port into master.